Make overview page update cluster states asynchronously
