+++ kqueen_ui/blueprints/ui/templates/ui/partial/_cluster_row.html
@@ -1,0 +1,43 @@
+{% from "ui/partial/tableaction.html" import render_cluster_row_actions with context %}
+
+{% macro render_cluster_row(cluster, index) %}
+        <tr {% if cluster.state in config.CLUSTER_TRANSIENT_STATES %} class="in-transition" data-href="{{ url_for('ui.cluster_row', cluster_id=cluster.id, index=index)  }}"{% endif %} data-index="{{ index }}" data-status="{{ cluster.state }}">
+          <td class="table-col-4">
+            {% if is_authorized(session, 'cluster:get', cluster) %}
+            <a href="{{ url_for('ui.cluster_detail', cluster_id=cluster.id) }}">{{ cluster.name }}</a>
+            {% else %}
+            {{ cluster.name }}
+            {% endif %}
+          </td>
+          <td class="table-col-4">{{ cluster.provisioner.name }}</td>
+          <td class="table-col-4">{{ cluster.created_at }}</td>
+          <td class="status_column">
+            {{ cluster.state|cluster_status_icon|safe }}
+          </td>
+          <td class="action_column">
+            {% if cluster.metadata %}
+            <a data-toggle="collapse" data-target="#cluster-{{ index }}" class="clickable" title="More informations">
+              <i class="mdi mdi-information-outline"></i>
+            </a>
+            {% endif %}
+            {{ render_cluster_row_actions(cluster) }}
+          </td>
+        </tr>
+        {% if cluster.metadata %}
+        <tr id="metadata-cluster-{{ index }}">
+          <td colspan="5" style="padding: 0px; border-top: 0px;">
+            <div style="margin: 8px;" id="cluster-{{ index }}" class="collapse">
+              <table class="table table-hover no-actions" id="table-clustermeta">
+                <tbody>
+                  {% for key, value in cluster.metadata.items() %}
+                  <tr>
+                    <th style="width: 300px;">{{ key }}</th>
+                    <td>{{ value }}</td>
+                  </tr>
+                  {% endfor %}
+                </tbody>
+              </table>
+            </div>
+          </td>
+        {% endif %}
+{% endmacro %}
