+++ kqueen_ui/blueprints/ui/templates/ui/partial/cluster_row.html
@@ -1,0 +1,3 @@
+{% from "ui/partial/_cluster_row.html" import render_cluster_row with context %}
+
+{{ render_cluster_row(cluster, index) }}
