+++ kqueen_ui/blueprints/ui/templates/ui/partial/overview_pies.html
@@ -1,0 +1,3 @@
+{% from "ui/partial/_overview_pies.html" import render_overview_pies with context %}
+
+{{ render_overview_pies(overview) }}
