+++ kqueen_ui/blueprints/ui/test_utils.py
@@ -1,5 +1,6 @@
+from kqueen_ui.api import KQueenResponse
 from kqueen_ui.conftest import cluster_status
-from .utils import status_for_cluster_detail
+from .utils import sanitize_resource_metadata, status_for_cluster_detail
 
 import pytest
 
@@ -146,3 +147,20 @@ class TestEmptyStatusForCluster:
     @pytest.mark.parametrize('key', ['services', 'addons', 'nodes', 'deployments', 'overview'])
     def test_status_for_cluster_detail_empty(self, key):
         assert key in self.status
+
+
+def test_sanitize_resource_metadata(app, user, cluster, provisioner, provisioner_engines, monkeypatch):
+    def mock_engines(self):
+        response = KQueenResponse()
+        response.data = provisioner_engines
+        return response
+    monkeypatch.setattr('kqueen_ui.api.ProvisionerManager.engines', mock_engines)
+    session = {'user': user}
+    parsed_clusters, parsed_provisioners, overview = sanitize_resource_metadata(session, [cluster], [provisioner])
+    parsed_cluster_metadata = parsed_clusters[0]['metadata']
+    parsed_provisioner_parameters = parsed_provisioners[0]['parameters']
+    overview_keys = ['cluster_count', 'cluster_max', 'cluster_health', 'provisioner_count', 'provisioner_max', 'provisioner_health']
+    assert parsed_cluster_metadata == cluster['metadata']
+    assert parsed_provisioner_parameters['username'] == provisioner['parameters']['username']
+    assert parsed_provisioner_parameters['password'] == '*****************'
+    assert all(key in overview for key in overview_keys)
